# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KepRoop API Documentation",
    "version": "1.0"
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    },
    {
      "CookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Prefix with 'Api-Key ' or 'Bearer kp_'"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Browser Session Cookie (Automatically used if logged in)"
      }
    }
  },
  "paths": {
...
    "/api/albums/{id}/activity": {
      "get": {
        "tags": [
          "Albums"
        ],
        "summary": "Get album activity",
        "description": "Get activity logs for an album.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity logs"
          }
        }
      }
    },
    "/api/albums/{id}/favorite": {
      "post": {
        "tags": [
          "Albums"
        ],
        "summary": "Toggle favorite",
        "description": "Add or remove album from favorites.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite status toggled"
          }
        }
      }
    },
    "/api/albums/{id}/folders/{folderId}": {
      "put": {
        "tags": [
          "Albums"
        ],
        "summary": "Rename folder",
        "description": "Rename a folder.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "folderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder renamed"
          }
        }
      },
      "delete": {
        "tags": [
          "Albums"
        ],
        "summary": "Delete folder",
        "description": "Delete a folder.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "folderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder deleted"
          }
        }
      }
    },
    "/api/albums/{id}/folders": {
      "get": {
        "tags": [
          "Albums"
        ],
        "summary": "List folders",
        "description": "List all folders in an album.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of folders"
          }
        }
      },
      "post": {
        "tags": [
          "Albums"
        ],
        "summary": "Create folder",
        "description": "Create a new folder in an album.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder created"
          }
        }
      }
    },
...
  },
...
}